EES-4705 Renaming the `Views` folder to `ViewModels` and updating namespaces and imports
