+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Api/Services/Interfaces/IPublicationService.cs
@@ -1,5 +1,5 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Api.Views;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Api.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Api.Services.Interfaces;
