+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Api/ViewModels/PublicationViewModels.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Api.Views;
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Api.ViewModels;
 
 public record PublicationListViewModel(Guid Id, string Title, string Slug, string Summary, DateTimeOffset LastPublished);
